requirements.txt
streamlit==1.42.0
google-genai>=0.5.0
langgraph>=0.2.70
tavily-python>=0.5.0
pydantic>=2.9.0
python-dotenv>=1.0.1
rich>=13.9.4
